:hammer: [Refactor] Solution 폴더 삭제
