+++ src/main/java/com/simter/domain/calendar/converter/CalendarsConverter.java
@@ -8,9 +8,9 @@ import com.simter.domain.calendar.dto.CalendarsResponseDto.CalendarsHomeDayDto;
 import com.simter.domain.calendar.dto.CalendarsResponseDto.NewCalendarsDto;
 import com.simter.domain.calendar.entity.Calendars;
 import com.simter.domain.chatbot.entity.CounselingLog;
+import com.simter.domain.chatbot.entity.Solution;
 import com.simter.domain.chatbot.repository.CalendarCounselingLogRepository;
 import com.simter.domain.member.entity.Member;
-import com.simter.domain.solution.entity.Solution;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
