+++ src/main/java/com/simter/domain/calendar/service/CalendarsService.java
@@ -14,11 +14,11 @@ import com.simter.domain.calendar.dto.CalendarsResponseDto.NewCalendarsDto;
 import com.simter.domain.calendar.entity.Calendars;
 import com.simter.domain.calendar.repository.CalendarsRepository;
 import com.simter.domain.chatbot.entity.CounselingLog;
+import com.simter.domain.chatbot.entity.Solution;
 import com.simter.domain.chatbot.repository.CounselingLogRepository;
+import com.simter.domain.chatbot.repository.SolutionRepository;
 import com.simter.domain.member.entity.Member;
 import com.simter.domain.member.repository.MemberRepository;
-import com.simter.domain.solution.entity.Solution;
-import com.simter.domain.solution.repository.SolutionRepository;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
